backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
flask-cors==4.0.0
# Para testes:
pytest==7.4.3
pytest-cov==4.1.0

